requirements.txt
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pdfkit==1.0.0
pdftotext==2.2.1
psycopg2==2.9.1
PyJWT==2.3.0
PyPDF2==1.26.0
python-dotenv==0.19.1
redis==4.0.2
Werkzeug==2.0.2
wkhtmltopdf==0.2
wrapt==1.13.3
